♻️ refactor: 중복 함수 utils로 분리
